requirements.txt
asgiref==3.6.0
Django==3.2.18
django-simple-history==3.2.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
